rename some files inside vec/sink dir
