+++ be/src/exec/data_sink.cpp
@@ -34,10 +34,10 @@
 #include "runtime/result_file_sink.h"
 #include "runtime/result_sink.h"
 #include "runtime/runtime_state.h"
-#include "vec/sink/result_sink.h"
 #include "vec/sink/vdata_stream_sender.h"
 #include "vec/sink/vmysql_table_sink.h"
 #include "vec/sink/vresult_file_sink.h"
+#include "vec/sink/vresult_sink.h"
 #include "vec/sink/vtablet_sink.h"
 
 namespace doris {
