+++ be/src/vec/CMakeLists.txt
@@ -194,8 +194,8 @@ set(VEC_FILES
   olap/vcollect_iterator.cpp
   olap/block_reader.cpp
   olap/olap_data_convertor.cpp
-  sink/mysql_result_writer.cpp
-  sink/result_sink.cpp
+  sink/vmysql_result_writer.cpp
+  sink/vresult_sink.cpp
   sink/vdata_stream_sender.cpp
   sink/vtablet_sink.cpp
   sink/vmysql_table_writer.cpp
