+++ be/src/vec/sink/vmysql_result_writer.h
@@ -20,7 +20,7 @@
 #include "util/mysql_row_buffer.h"
 #include "util/runtime_profile.h"
 #include "vec/core/block.h"
-#include "vec/sink/result_writer.h"
+#include "vec/sink/vresult_writer.h"
 
 namespace doris {
 class BufferControlBlock;
